pyproject.toml
[tool.poetry]
name = "ebs-python-internship-test"
version = "0.1.0"
description = ""
authors = ["korsarking <korsarkings@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
attrs = "^23.1.0"
asgiref = "^3.7.2"
certifi = "^2023.7.22"
coreapi = "^2.3.3"
django = "^4.2.6"
dj-email-url = "^1.0.6"
django-cache-url = "^3.4.4"
django-cors-headers = "^4.2.0"
django-filter = "^23.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-redis = "^5.4.0"
drf-util = "^1.0.17"
environs = "^9.5.0"
factory-boy = "^3.3.0"
idna = "^3.4"
jsonschema = ">=4.10.0,<4.18.0"
openapi-codec = "^1.3.2"
packaging = "^23.2"
psycopg = "^3.1.12"
psycopg2-binary = "^2.9.9"
python = "^3.11"
pytz = "^2023.3.post1"
randomtimestamp = "^2.2"
rapidfuzz = ">=2.2.0,<3.0.0"
redis = "^5.0.1"
requests = "^2.31.0"
requests-toolbelt = "^1.0.0"
ruamel-yaml = "^0.17.35"
ruamel-yaml-clib = "^0.2.8"
setuptools = "^68.2.2"
shellingham = "^1.5.3"
six = "^1.16.0"
sqlparse = "^0.4.4"
swagger-spec-validator = "^3.0.3"
tomlkit = "^0.12.1"
traceback2 = "^1.4.0"
trove-classifiers = "^2023.9.19"
typing-extensions = "^4.8.0"
tzdata = "^2023.3"
uritemplate = "^4.1.1"
urllib3 = "^2.0.6"
varname = "^0.12.0"
wheel = "^0.41.2"
zipp = "^3.17.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
